Player can lose hp

+++ Bulbo/Assets/Scripts/Enemies/EnemyAttributeSO.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "EnemyAttributeSO", menuName = "Scriptable Objects/EnemyAttributeSO")]
 public class EnemyAttributeSO : ScriptableObject
 {
-    public float health = 100;
-    public float speed = 20;
-    public float attackDamage = 5;
-    public float xpWorth = 5;
+    public int health = 100;
+    public int speed = 20;
+    public int attackDamage = 5;
+    public int xpWorth = 5;
 }
